Array length (now an uint) should use public storage too (#88)

+++ test/libsolidity/semanticTests/array/shielded_array_storage_push_pop.sol
@@ -5,7 +5,7 @@ contract C {
             storageArray.push();
         while(storageArray.length > 0)
             storageArray.pop();
-        return uint(storageArray.length);
+        return storageArray.length;
     }
 }
 // ----
